requirements.dev.txt
bkcharts==0.2
bokeh==0.12.7
certifi==2016.2.28
click==6.7
cycler==0.10.0
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
opencv-python==3.3.0.10
pandas==0.20.3
Pillow==4.2.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
six==1.10.0
tornado==4.5.2
Werkzeug==0.12.2
wincertstore==0.2
